+++ paddle_billing/Resources/Transactions/Operations/__init__.py
@@ -2,6 +2,7 @@ from paddle_billing.Resources.Transactions.Operations.CreateTransaction import C
 from paddle_billing.Resources.Transactions.Operations.ListTransactions import ListTransactions
 from paddle_billing.Resources.Transactions.Operations.List.Includes import Includes as TransactionIncludes
 from paddle_billing.Resources.Transactions.Operations.List.Origin import Origin as TransactionOrigin
+from paddle_billing.Resources.Transactions.Operations.PreviewTransaction import PreviewTransaction
 from paddle_billing.Resources.Transactions.Operations.PreviewTransactionByAddress import PreviewTransactionByAddress
 from paddle_billing.Resources.Transactions.Operations.PreviewTransactionByCustomer import PreviewTransactionByCustomer
 from paddle_billing.Resources.Transactions.Operations.PreviewTransactionByIP import PreviewTransactionByIP