+++ paddle_billing/Resources/Transactions/TransactionsClient.py
@@ -11,6 +11,7 @@ from paddle_billing.Resources.Transactions.Operations import (
     CreateTransaction,
     ListTransactions,
     UpdateTransaction,
+    PreviewTransaction,
     PreviewTransactionByAddress,
     PreviewTransactionByCustomer,
     PreviewTransactionByIP,
@@ -80,7 +81,10 @@ class TransactionsClient:
         return Transaction.from_dict(parser.get_data())
 
     def preview(
-        self, operation: PreviewTransactionByAddress | PreviewTransactionByCustomer | PreviewTransactionByIP
+        self,
+        operation: (
+            PreviewTransaction | PreviewTransactionByAddress | PreviewTransactionByCustomer | PreviewTransactionByIP
+        ),
     ) -> TransactionPreview:
         self.response = self.client.post_raw("/transactions/preview", operation)
         parser = ResponseParser(self.response)